Added `BurtSocket.tryHandshake()` (#15)
There's scenarios where you might want to listen to the incoming
heartbeats or settings updates. For example, in the dashboard it listens
to incoming UpdateSetting messages for changing modes.

Since these messages are never added to the stream, certain handshakes
will never be completed since they are never added to the stream. This commit fixes that. 

This also introduces `tryHandshake()`, which handles the boilerplate of receiving an echo response.

---------

Co-authored-by: Levi Lesches <levilesches@gmail.com>